refactor : 가게 존재 검증 및 삭제 확인 메서드 통일화
가게 검증할 때 각 서비스마다 검증 로직 짜야 되는 상황 -> StoreService만 추가하여 처리하는 거로 변경

의존 클래서 4개 -> 1개로 축소
