+++ src/main/java/org/baratie/yumyum/domain/favorite/controller/FavoriteController.java
@@ -4,7 +4,6 @@ import lombok.RequiredArgsConstructor;
 import org.baratie.yumyum.domain.favorite.dto.FavoriteDto;
 import org.baratie.yumyum.domain.favorite.service.FavoriteService;
 import org.baratie.yumyum.domain.member.domain.CustomUserDetails;
-import org.baratie.yumyum.domain.member.domain.Member;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
