+++ src/main/java/org/baratie/yumyum/domain/favorite/service/FavoriteService.java
@@ -8,22 +8,26 @@ import org.baratie.yumyum.domain.member.domain.CustomUserDetails;
 import org.baratie.yumyum.domain.member.domain.Member;
 import org.baratie.yumyum.domain.member.repository.MemberRepository;
 import org.baratie.yumyum.domain.store.domain.Store;
-import org.baratie.yumyum.domain.store.exception.StoreNotFoundException;
-import org.baratie.yumyum.domain.store.repository.StoreRepository;
-import org.baratie.yumyum.global.exception.ErrorCode;
+import org.baratie.yumyum.domain.store.service.StoreService;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class FavoriteService {
 
-    private final StoreRepository storeRepository;
     private final FavoriteRepository favoriteRepository;
     private final MemberRepository memberRepository;
+    private final StoreService storeService;
 
+    /**
+     * 즐겨찾기 추가 or 취소
+     * @param customUserDetails 인증된 사용자 객체
+     * @param request 즐겨찾기 추가할 가게 id와 즐겨찾기 상태
+     */
     public void addFavorite(CustomUserDetails customUserDetails, FavoriteDto request) {
 
-        Store store = validationStoreId(request.getStoreId());
+        Store store = storeService.validationStoreId(request.getStoreId());
+
         Member member = memberRepository.findById(customUserDetails.getId()).orElseThrow();
 
         favoriteRepository.exist(customUserDetails.getId(), store.getId()).ifPresentOrElse(
@@ -34,16 +38,5 @@ public class FavoriteService {
                 () -> favoriteRepository.save(Favorite.insertFavorite(member, store))
         );
 
-    }
-
-    /**
-     * 가게가 DB에 존재하는지 확인
-     * @param storeId 가게 pk
-     * @return Store
-     */
-    private Store validationStoreId(Long storeId) {
-        return storeRepository.findById(storeId).orElseThrow(
-                () -> new StoreNotFoundException(ErrorCode.STORE_NOT_FOUND)
-        );
     }
 }
