+++ src/main/java/org/baratie/yumyum/domain/store/repository/StoreCustomRepository.java
@@ -1,0 +1,13 @@
+package org.baratie.yumyum.domain.store.repository;
+
+import org.baratie.yumyum.domain.store.domain.Store;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface StoreCustomRepository {
+
+    Optional<Store> existAndDeletedStore(Long storeId);
+
+}
