+++ src/main/java/org/baratie/yumyum/domain/store/repository/StoreRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface StoreRepository extends JpaRepository<Store, Long> {
+public interface StoreRepository extends JpaRepository<Store, Long>, StoreCustomRepository{
 
     boolean existsByName(String storeName);
 }
