+++ src/main/java/org/baratie/yumyum/domain/store/repository/impl/StoreRepositoryImpl.java
@@ -1,0 +1,40 @@
+package org.baratie.yumyum.domain.store.repository.impl;
+
+import com.querydsl.core.types.dsl.BooleanExpression;
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.RequiredArgsConstructor;
+import org.baratie.yumyum.domain.store.domain.Store;
+import org.baratie.yumyum.domain.store.repository.StoreCustomRepository;
+
+import java.util.Optional;
+
+import static org.baratie.yumyum.domain.store.domain.QStore.*;
+
+@RequiredArgsConstructor
+public class StoreRepositoryImpl implements StoreCustomRepository {
+
+    private final JPAQueryFactory query;
+
+    /**
+     * @param storeId 가게가 존재하는지 검증이 필요한 가게 ID
+     * @return Optional<Store>
+     */
+    @Override
+    public Optional<Store> existAndDeletedStore(Long storeId) {
+
+        Store findStore = query.selectFrom(store)
+                .where(storeIdEq(storeId).and(store.isClosed.isFalse()))
+                .fetchOne();
+
+        return Optional.ofNullable(findStore);
+    }
+
+    /**
+     * 가게가 있는지 확인
+     * @param storeId 조회할 가게 id
+     * @return
+     */
+    private BooleanExpression storeIdEq(Long storeId) {
+        return store.id.eq(storeId);
+    }
+}
